+++ src/ui/material/toggle/src/toggle.module.ts
@@ -5,10 +5,10 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { FormlyMatFormFieldModule } from '@ngx-formly/material/form-field';
 
-import { FormlyToggleTypeComponent } from './toggle.type';
+import { FormlyFieldToggle } from './toggle.type';
 
 @NgModule({
-  declarations: [FormlyToggleTypeComponent],
+  declarations: [FormlyFieldToggle],
   imports: [
     CommonModule,
     ReactiveFormsModule,
@@ -18,7 +18,7 @@ import { FormlyToggleTypeComponent } from './toggle.type';
       types: [
         {
           name: 'toggle',
-          component: FormlyToggleTypeComponent,
+          component: FormlyFieldToggle,
           wrappers: ['form-field'],
         },
       ],