+++ src/ui/material/toggle/src/toggle.type.ts
@@ -19,7 +19,7 @@ import { MatSlideToggle } from '@angular/material/slide-toggle';
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class FormlyToggleTypeComponent extends FieldType {
+export class FormlyFieldToggle extends FieldType {
   @ViewChild(MatSlideToggle, { static: true }) slideToggle!: MatSlideToggle;
   defaultOptions = {
     templateOptions: {